+++ packages/components/progress-bar/src/elements/label.tsx
@@ -1,9 +1,13 @@
 // @ts-nocheck
 import React from "react"
 import { isUndefined, isEmpty } from "@wpmudev/sui-utils"
+import { ProgressBarProps } from "../progress-bar.types"
 
 // Build "label" element
-const Label = ({ className, children }) => {
+const Label: React.FC<Pick<ProgressBarProps, "className" | "children">> = ({
+	className,
+	children,
+}) => {
 	const has = {}
 	const set = {}