+++ packages/components/progress-bar/src/progress-bar.tsx
@@ -1,6 +1,6 @@
 // @ts-nocheck
 import React from "react"
-import { isFunction, generateCN } from "@wpmudev/sui-utils"
+import { generateCN } from "@wpmudev/sui-utils"
 import { Button } from "@wpmudev/sui-button"
 
 // Import required components