+++ packages/components/rich-text-editor/__tests__/rich-test-editor.test.js
@@ -1,7 +1,5 @@
 "use strict"
 
-const { Selector } = require("../src")
-
 describe("@wpmudev/sui-selector", () => {
-	it("needs tests")
+	// Tests to be added
 })