+++ packages/components/select/src/variants/select-search.tsx
@@ -1,11 +1,11 @@
 import React from "react"
 
 // Import required component(s).
-import { Select as Base } from "./select-base"
+import { Select as Base, SelectBaseProps } from "./select-base"
 
 // Build "Search Select" component.
-const SearchSelect = ({ ...props }) => {
-	return <Base id={props?.id} options={props?.options ?? []} {...props} />
+const SearchSelect: React.FC<SelectBaseProps> = ({ ...props }) => {
+	return <Base {...props} isSearchable={true} />
 }
 
 // Publish required component(s).