+++ packages/components/accordion/CHANGELOG.md
@@ -1,11 +1,24 @@
 # Change Log
 
-All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org/) for commit guidelines.
+## 0.0.1
+
+### Patch Changes
+
+- Initial release
+- Updated dependencies
+  - @wpmudev/sui-icons@0.0.1
+  - @wpmudev/sui-checkbox@0.0.1
+  - @wpmudev/sui-utils@0.0.1
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org/) for commit guidelines.
 
 ## Unreleased (2023-03-20)
 
 #### ✨ New Features
+
 - feat: Initial release ([@creador-dev](https://github.com/creador-dev))
 
 #### Committers: 1
+
 - Pawan Kumar ([@creador-dev](https://github.com/creador-dev))